# [WIP] Implement boolean preambles