I have been recently playing with idea of automatic preamble calculation for `(or ...)`, `(and ...)` and similar queries.

If I understand correctly, preamble is dropped on master for any boolean expression, except `(and ...)`. However, `(or ...)` could also be optimised if all the sexps inside `or` have non-nil predicates.

This pull request is a not-yet-ready-to-merge draft where I want to ask your opinion about this idea. I implemented boolean expressions `(and ...)`, `(or ...)`, `(when ...)`, `(unless ...)`, and `(not ...)` as actual predicates, so that  preambles can be transparently defined. 

Some benchmarks with `elp` on my setup with complex agenda queries:

;;Function  number_of_invocations  total_time average_time

;; Old (agenda n)
;;org-agenda               1           45.655149553  45.655149553
;;org-ql--select           2255        37.698226351  0.0167176170
;;re-search-forward        336625      6.2799195780  1.865...e-05
;;org-ql--query-preamble   2265        0.0108973820  4.811...e-06
;;org-ql--normalize-query  2265        0.0088714750  3.916...e-06

;; New (agenda n)
;;org-agenda               1           22.979330424  22.979330424
;;org-ql--select           1261        15.829318511  0.0125529885
;;re-search-forward        256703      4.8654712150  1.895...e-05
;;org-ql--normalize-query  1274        0.0073856340  5.797...e-06
;;org-ql--query-preamble   1274        0.0045016170  3.533...e-06

Note the difference in number of `re-search-forward` calls: 256703 (new) vs. 336625 (old). This is mostly because `(or ...)` queries can now have preamble. I have seen up to several times less calls depending on what kind of query I am using.

The caveat is (sometimes) increased time spend on each `re-search-forward`. Combined preamble for `(or ...)` consists of `(rx `(| individual-preambles))` increasing the complexity of preamble regexp. For very small queries involving many timestamps, regexp search sometimes ate up all the speed improvement as the time for each individual regexp search increased several times.