@yantar92 2021-04-05
Note: While looking into this subject, I noticed that `tags-local` predicate is inefficient if its preamble is unused (i.e. `(or (tags-local "TAG") ...`). `tags-local` body calls slow `org-ql--tags-at`.