@alphapapa 2021-04-05
This is very interesting, thanks.  I hadn't thought of using the new defpred macro to define normalizer- and preamble-only versions of the booleans.

Assuming that everything works as intended, and we don't cause any edge cases to regress in performance, this is certainly an improvement I'd like to make.  Here are a few thoughts from a quick look:

1.  I see you're calling a function, `rec`, which is defined with `cl-labels` in another function.  Although it may work, I think it's at least an unintended result of the implementation (since `cl-labels` temporarily rebinds symbol functions), if not a violation of lexical scope (again, perhaps in spirit if not technically, in Elisp).  In any case, I think it shouldn't be done that way, and it makes the code hard to understand (i.e. where the `rec` being called is defined).
2.  In that vein, I see some lexical binding-related warnings in the tests.  Sometimes those do indicate actual problems, especially possible in light of the previous point.
3.  The tests also are not passing, but I'm not sure if it's due to this PR, or if that dash-functional obsolescence warning is causing it (it shouldn't, but I haven't checked).
4. You mentioned optimizing timestamp regexps.  Not that this would obviate this PR, but FYI, I have a branch that does that explicitly, and it works pretty well.  See https://github.com/alphapapa/org-ql/tree/wip/ts-optimized-regexps  I haven't had time to work on it for a little while, but I think it just needs a little more polishing and testing (I think I have some notes about it in the `meta/notes` branch).

> Note: While looking into this subject, I noticed that tags-local predicate is inefficient if its preamble is unused (i.e. (or (tags-local "TAG") ...). tags-local body calls slow org-ql--tags-at.

I'm not sure what you mean.  If its preamble is unused, what else would it do?  And `org-ql--tags-at` is a caching function, which speeds up future local-tags queries on headings.

Thanks for your work on this.