@yantar92 2021-04-05
> I see you're calling a function, rec, which is defined with cl-labels in another function. 

I agree that it is difficult to read. But what would be an alternative to get preamble/normalization recursively? The best way I can see is documenting it as you did for `,predicate-names`. Also, existing org-ql code calls `rec` directly in `ancestors` and `parent` predicates.

> In that vein, I see some lexical binding-related warnings in the tests.

Those are generated by expansion of `org-ql-defpred` macro. Without :body, predicates generate `(cl-defun ,fn-name ,args ,docstring nil)`. Of course, arguments are unused there.

> The tests also are not passing

It is related to another idea I have not mentioned here. 

I wanted to make use of predicates :body code moving point. For now, if predicate code moves the point, org-ql continues the search from the heading next to the new position. This might be useful for additional speedup. Say, outline-path predicate could search forward for regexp `(rx '(or ,@strings))` if there is no match at point and otherwise let org-ql continue from next heading.

With current `org-ql` code the idea would not work though. If outline-path moves point, query like `(not (outline-path...))` would give wrong result. So, I added `save-excursion` around predicates that must not affect point position. The result is failing tests because normalized queries do not match the expected result in tests.

Anyway, It is not reletad to this particular pull request and I will remove `save-excursion` code later.

> See https://github.com/alphapapa/org-ql/tree/wip/ts-optimized-regexps

Intresting. Thanks for the pointer.

> I'm not sure what you mean. If its preamble is unused, what else would it do? And org-ql--tags-at is a caching function, which speeds up future local-tags queries on headings.

`org-ql--tags-at` calculates inherited tags unconditionally. Inherited tags are not needed for `tags-local` preamble and, in fact, created bottleneck in one of my queries where I explicitly wanted to use `tags-local` for performance reasons.